Feat: [SpecSens] Reshape freq-div output to time-div
To leverage the multi-thread model in savannah, we need to partition the tasks
into frequency (or subcarrier) division. After scheduler update, DumpToFile()
cannot generate the same output format as it takes some re-arranging and file
open/close time.

svnh_reshape_dumpped_iq.py helps transform the new format (one file represents a
subband within a time period, i.e. ping-pong buffer range) back to old format
(each file is one symbol of full band), so that other scripts can be used as is.

svnh_concatenate_dumpped_iq.sh is deprecated. When DumpToFile() tries to dump
to more files (so that each file is a subband of a symbol), it takes too much
time. svnh_concatenate_dumpped_iq.sh is used to combine these many files back
to the old format. As dumping this way is proved to be too slow (and will mess
up the ping-pong buffer), the script is not useful.